Default to internal-only traffic
The use case for this proxy is likely to be within the same cloud
project, so this seems like a better default and is more secure. The
quick setup guide already makes opinionated decisions (like naming) and
creates everything in one cloud project, so someone who wanted to
diverge from this would already need to modify other steps as well.

Signed-off-by: Geoffrey Martin-Noble <gcmn@google.com>